Check for a parameter with nullable() in migrations
If you pass `false` into nullable(), for example when turning a column
from nullable to non-null, the migration parsing code would ignore this
and assume the column was nullable, causing false "potentially null
value" errors when the column was used.

This commit adds a check for either no parameter or a true parameter
before making the column as nullable.

Fixes psalm/psalm-plugin-laravel#274