Add special `dev` to composer keywords
The special dev keyword will warn users who run composer require psalm/plugin-laravel with the following prompt:
> composer require psalm/plugin-laravel
The package you required is recommended to be placed in require-dev (because it is tagged as "dev") but you did not use --dev.
Do you want to re-run the command with --dev? [yes]?